Move the current Validator implementation within the `scylla` package
