+++ migrator/src/main/scala/com/scylladb/migrator/scylla/ScyllaValidator.scala
@@ -1,0 +1,134 @@
+package com.scylladb.migrator.scylla
+
+import com.datastax.spark.connector.{
+  toSparkContextFunctions,
+  ColumnName,
+  SomeColumns,
+  TTL,
+  WriteTime
+}
+import com.datastax.spark.connector.cql.{ CassandraConnector, Schema }
+import com.scylladb.migrator.Connectors
+import com.scylladb.migrator.config.{ MigratorConfig, SourceSettings, TargetSettings }
+import com.scylladb.migrator.validation.RowComparisonFailure
+import org.apache.log4j.LogManager
+import org.apache.spark.sql.SparkSession
+import com.datastax.oss.driver.api.core.ConsistencyLevel
+import com.datastax.spark.connector.rdd.ReadConf
+
+/** The C* to Scylla migration validator */
+object ScyllaValidator {
+
+  private val log = LogManager.getLogger("com.scylladb.migrator.scylla")
+
+  def runValidation(
+    sourceSettings: SourceSettings.Cassandra,
+    targetSettings: TargetSettings.Scylla,
+    config: MigratorConfig)(implicit spark: SparkSession): List[RowComparisonFailure] = {
+    val sourceConnector: CassandraConnector =
+      Connectors.sourceConnector(spark.sparkContext.getConf, sourceSettings)
+    val targetConnector: CassandraConnector =
+      Connectors.targetConnector(spark.sparkContext.getConf, targetSettings)
+
+    val renameMap = config.renames.map(rename => rename.from -> rename.to).toMap
+    val sourceTableDef =
+      sourceConnector.withSessionDo(
+        Schema.tableFromCassandra(_, sourceSettings.keyspace, sourceSettings.table))
+
+    val source = {
+      val regularColumnsProjection =
+        sourceTableDef.regularColumns.flatMap { colDef =>
+          val alias = renameMap.getOrElse(colDef.columnName, colDef.columnName)
+
+          if (sourceSettings.preserveTimestamps)
+            List(
+              ColumnName(colDef.columnName, Some(alias)),
+              WriteTime(colDef.columnName, Some(alias + "_writetime")),
+              TTL(colDef.columnName, Some(alias + "_ttl"))
+            )
+          else List(ColumnName(colDef.columnName, Some(alias)))
+        }
+
+      val primaryKeyProjection =
+        (sourceTableDef.partitionKey ++ sourceTableDef.clusteringColumns)
+          .map(colDef => ColumnName(colDef.columnName, renameMap.get(colDef.columnName)))
+
+      val consistencyLevel = sourceSettings.consistencyLevel match {
+        case "LOCAL_QUORUM" => ConsistencyLevel.LOCAL_QUORUM
+        case "QUORUM"       => ConsistencyLevel.QUORUM
+        case "LOCAL_ONE"    => ConsistencyLevel.LOCAL_ONE
+        case "ONE"          => ConsistencyLevel.ONE
+        case _              => ConsistencyLevel.LOCAL_QUORUM
+      }
+      if (consistencyLevel.toString == sourceSettings.consistencyLevel) {
+        log.info(
+          s"Using consistencyLevel [${consistencyLevel}] for VALIDATOR SOURCE based on validator source config [${sourceSettings.consistencyLevel}]")
+      } else {
+        log.info(
+          s"Using DEFAULT consistencyLevel [${consistencyLevel}] for VALIDATOR SOURCE based on unrecognized validator source config [${sourceSettings.consistencyLevel}]")
+      }
+
+      spark.sparkContext
+        .cassandraTable(sourceSettings.keyspace, sourceSettings.table)
+        .withConnector(sourceConnector)
+        .withReadConf(
+          ReadConf
+            .fromSparkConf(spark.sparkContext.getConf)
+            .copy(
+              splitCount       = sourceSettings.splitCount,
+              fetchSizeInRows  = sourceSettings.fetchSize,
+              consistencyLevel = consistencyLevel
+            )
+        )
+        .select(primaryKeyProjection ++ regularColumnsProjection: _*)
+    }
+
+    val joined = {
+      val regularColumnsProjection =
+        sourceTableDef.regularColumns.flatMap { colDef =>
+          val renamedColName = renameMap.getOrElse(colDef.columnName, colDef.columnName)
+
+          if (sourceSettings.preserveTimestamps)
+            List(
+              ColumnName(renamedColName),
+              WriteTime(renamedColName, Some(renamedColName + "_writetime")),
+              TTL(renamedColName, Some(renamedColName + "_ttl"))
+            )
+          else List(ColumnName(renamedColName))
+        }
+
+      val primaryKeyProjection =
+        (sourceTableDef.partitionKey ++ sourceTableDef.clusteringColumns)
+          .map(colDef => ColumnName(renameMap.getOrElse(colDef.columnName, colDef.columnName)))
+
+      val joinKey = (sourceTableDef.partitionKey ++ sourceTableDef.clusteringColumns)
+        .map(colDef => ColumnName(renameMap.getOrElse(colDef.columnName, colDef.columnName)))
+
+      source
+        .leftJoinWithCassandraTable(
+          targetSettings.keyspace,
+          targetSettings.table,
+          SomeColumns(primaryKeyProjection ++ regularColumnsProjection: _*),
+          SomeColumns(joinKey: _*))
+        .withConnector(targetConnector)
+    }
+
+    joined
+      .flatMap {
+        case (l, r) =>
+          RowComparisonFailure.compareRows(
+            l,
+            r,
+            config.validation.floatingPointTolerance,
+            config.validation.timestampMsTolerance,
+            config.validation.ttlToleranceMillis,
+            config.validation.writetimeToleranceMillis,
+            config.validation.compareTimestamps
+          )
+      }
+      .take(config.validation.failuresToFetch)
+      .toList
+
+  }
+
+}
